requirements.txt
yt-dlp
moviepy
Pillow
whisper
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
